+++ src/test/java/kitchenpos/RestAssuredFixture.java
@@ -1,0 +1,62 @@
+package kitchenpos;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class RestAssuredFixture {
+
+    public static void 성공_200_OK(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    public static void 생성됨_201_CREATED(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    public static void 콘텐츠_없음_204_NO_CONTENT(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    public static ExtractableResponse<Response> 생성_요청(String path, Object params) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().post(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 목록_조회_요청(String path) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().get(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 수정_요청(String path, Object params) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().put(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 삭제_요청(String path) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().delete(path)
+                .then().log().all()
+                .extract();
+    }
+}
