--- src/test/java/kitchenpos/fixture/MenuProductFixture.java
@@ -1,15 +1,0 @@
-package kitchenpos.fixture;
-
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.product.domain.Product;
-
-public class MenuProductFixture {
-
-    private MenuProductFixture() {
-        throw new UnsupportedOperationException();
-    }
-
-    public static MenuProduct create(Long seq, Product product, long quantity) {
-        return MenuProduct.of(seq, null, product, quantity);
-    }
-}
